+++ stream_alert_cli/utils.py
@@ -611,89 +611,3 @@ def _setup_threat_intel_downloader_subparser(subparser):
     _setup_threat_intel_auth_subparser(ti_subparsers)
 
 
-def _setup_rule_staging_subparser(subparser):
-    """Add the rule staging subparser: manage.py rule-staging [subcommand] [options]"""
-    rule_staging_subparsers = subparser.add_subparsers(dest="rule-staging subcommand",
-                                                       required=True)
-
-    _setup_rule_staging_enable_subparser(rule_staging_subparsers)
-    _setup_rule_staging_status_subparser(rule_staging_subparsers)
-    _setup_rule_staging_stage_subparser(rule_staging_subparsers)
-    _setup_rule_staging_unstage_subparser(rule_staging_subparsers)
-
-def _setup_rule_staging_enable_subparser(subparsers):
-    """Add the rule staging enable subparser: manage.py rule-staging enable"""
-    rule_staging_enable_parser = _generate_subparser(
-        subparsers,
-        'enable',
-        description='Enable or disable the rule staging feature',
-        subcommand=True
-    )
-
-    toggle_group = rule_staging_enable_parser.add_mutually_exclusive_group(required=True)
-    toggle_group.add_argument(
-        '-t',
-        '--true',
-        dest='enable',
-        help='Enable the rule staging feature',
-        action='store_true'
-    )
-    toggle_group.add_argument(
-        '-f',
-        '--false',
-        dest='enable',
-        help='Disable the rule staging feature',
-        action='store_false'
-    )
-
-
-def _setup_rule_staging_status_subparser(subparsers):
-    """Add the rule staging status subparser: manage.py rule-staging status"""
-    rule_staging_status_parser = _generate_subparser(
-        subparsers,
-        'status',
-        description='List all rules within the rule database and their staging status',
-        subcommand=True
-    )
-
-    rule_staging_status_parser.add_argument(
-        '-v',
-        '--verbose',
-        action='store_true',
-        help='Output additional information for rules in the database'
-    )
-
-
-def _setup_rule_staging_stage_subparser(subparsers):
-    """Add the rule staging stage subparser: manage.py rule-staging stage"""
-    rule_staging_stage_parser = _generate_subparser(
-        subparsers,
-        'stage',
-        description='Stage the provided rules',
-        subcommand=True
-    )
-
-    _add_default_rule_staging_args(rule_staging_stage_parser)
-
-
-def _setup_rule_staging_unstage_subparser(subparsers):
-    """Add the rule staging unstage subparser: manage.py rule-staging unstage"""
-    rule_staging_unstage_parser = _generate_subparser(
-        subparsers,
-        'unstage',
-        description='Unstage the provided rules',
-        subcommand=True
-    )
-
-    _add_default_rule_staging_args(rule_staging_unstage_parser)
-
-
-def _add_default_rule_staging_args(subparser):
-    """Add the default arguments to the rule staging parsers"""
-    subparser.add_argument(
-        'rules',
-        action=UniqueSetAction,
-        default=set(),
-        help='One or more rule to perform this action against, seperated by spaces',
-        nargs='+'
-    )
